Add explanation and improve BayesN
1. Add parameter descriptions
2. Add decimal_place option for probability in bayesian_network.ipynb

+++ aispace2/jupyter/bayes/bayes.py
@@ -22,7 +22,7 @@ class Displayable(StepDOMWidget):
     text_size = Integer(12).tag(sync=True)
     show_full_domain = Bool(False).tag(sync=True)
     detail_level = Integer(2).tag(sync=True)
-    decimal_place = Integer(10).tag(sync=True)
+    decimal_place = Integer(2).tag(sync=True)
 
     probCalculator = VE()
